reference/upskill-0.2.1/skills-test/pyproject.toml
[project]
name = "fast-agent-app"
version = "0.1.0"
description = "a simple fast-agent application"
readme = "README.md"
requires-python = "<3.14,>=3.13.5"
dependencies = [
  "beautifulsoup4>=4.14.2",
  "fast-agent-mcp==0.4.37",
  "huggingface-hub==1.3.2",
  "markdown-it-py>=4.0.0",
  "matplotlib>=3.10.7",
  "multilspy>=0.0.15",
  "python-dotenv>=1.2.1",
  "pyyaml>=6.0.3",
  "requests>=2.32.3",
]

[tool.uv]
package = true

# Optional convenience entry point if you later package this app:
[project.scripts]
fast-agent-app = "agent:main"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["agent"]
